+++ website/src/examples/mod.rs
@@ -12,6 +12,7 @@ mod basic_dropdown;
 mod basic_modal;
 mod basic_navbar;
 mod basic_pagination;
+mod basic_title;
 mod general_form;
 
 pub use basic_block::BasicBlock;
@@ -21,6 +22,7 @@ pub use basic_dropdown::BasicDropdown;
 pub use basic_modal::BasicModal;
 pub use basic_navbar::BasicNavbar;
 pub use basic_pagination::BasicPagination;
+pub use basic_title::BasicTitle;
 pub use general_form::GeneralForm;
 
 async fn get_code_example(name: &'static str) -> Result<String, JsValue> {