+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Checkbox Leptos component.
 - Block Leptos component.
+- Title and subtitle Leptos components.
 
 ## [0.0.0-alpha.9] - 2024-04-10